@@ -1,23 +1,22 @@
-import 'package:example/presenter.dart';
 import 'package:example/state.dart';
 import 'package:flutter/material.dart';
 import 'package:osam/osam.dart';
 
+import 'model.dart';
+
 void main() => runApp(MyApp());
 
-enum EventType { increment }
+enum EventType { increment, incrementValue }
 
 class MyApp extends StatelessWidget {
+  final store = Store<Counter>.single(Counter(), middleWares: []);
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       home: StoreProvider(
         child: MyHomePage(),
-        store: Store(states: [
-          Counter(),
-        ], middleWares: [
-          //    MyMiddleware()
-        ]),
+        store: store,
       ),
     );
   }
@@ -26,7 +25,7 @@ class MyApp extends StatelessWidget {
 class MyHomePage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final state = StoreProvider.of(context).getState<Counter>();
+    final state = StoreProvider.of(context).state as Counter;
     return Scaffold(
         body: Center(
           child: Column(
@@ -36,29 +35,58 @@ class MyHomePage extends StatelessWidget {
                 'You have pushed the button this many times:',
               ),
               StreamBuilder(
-                initialData: state.count,
-                stream: state.propertyStream<Counter, List<int>>((state) => state.count),
-                builder: (ctx, AsyncSnapshot<List<int>> snapshot) {
-                  return Text(
-                    snapshot.data.toString(),
-                    style: TextStyle(fontSize: 50),
+                initialData: state.count.values.toList(),
+                stream: state
+                    .propertyStream<Counter, List<Model>>((state) => state.count.values.toList()),
+                builder: (ctx, AsyncSnapshot<List<Model>> snapshot) {
+                  print('123');
+                  return Column(
+                    children: snapshot.data.map((model) => RText(model: model)).toList(),
                   );
                 },
               )
             ],
           ),
         ),
-        floatingActionButton:
-            PresenterProvider<MPresenter>(child: Button(), presenter: MPresenter()));
+        floatingActionButton: Button());
+  }
+}
+
+class RText extends StatelessWidget {
+  final Model model;
+  const RText({Key key, this.model}) : super(key: key);
+  @override
+  Widget build(BuildContext context) {
+    return StreamBuilder(
+      initialData: model.number,
+      stream: model.propertyStream<Model, int>((state) => state.number),
+      builder: (ctx, AsyncSnapshot<int> snapshot) {
+        print('model');
+        return GestureDetector(
+          onTap: () {
+            StoreProvider.of(context).dispatchEvent<Counter>(
+                event: Event.modify(
+                    reducerCaller: (state, bundle) => state.count[model.key]..increment()));
+          },
+          child: Text(
+            snapshot.data.toString(),
+            style: TextStyle(fontSize: 25),
+          ),
+        );
+      },
+    );
   }
 }
 
 class Button extends StatelessWidget {
+  int number = 0;
   @override
   Widget build(BuildContext context) {
     return FloatingActionButton(
       onPressed: () {
-        PresenterProvider.of<MPresenter>(context).increment();
+        number++;
+        StoreProvider.of(context).dispatchEvent<Counter>(
+            event: Event.modify(reducerCaller: (state, _) => state..increment(number)));
       },
       tooltip: 'Increment',
       child: Icon(Icons.add),
