@@ -1,18 +1,23 @@
-import 'package:example/state.dart';
-import 'package:osam/presentation/presenter.dart';
-import 'package:osam/util/event.dart';
-
-import 'main.dart';
-
-class MPresenter extends Presenter {
-  void increment() {
-    this.store.dispatchEvent<Counter>(
-        event: Event.modify(reducer: (state, _) => state.increment(1), type: EventType.increment));
-  }
-
-  @override
-  void dispose() {}
-
-  @override
-  void init() {}
-}
+//import 'package:example/state.dart';
+//import 'package:osam/presentation/presenter.dart';
+//import 'package:osam/util/event.dart';
+//
+//import 'main.dart';
+//
+//class MPresenter extends Presenter {
+//  void increment() {
+//    final targerState = this.store.
+//
+//
+//
+//    this.store.dispatchEvent<Counter>(
+//        event: Event.modify(
+//            reducer: (state, _) => state.increment(number), type: EventType.increment));
+//  }
+//
+//  @override
+//  void dispose() {}
+//
+//  @override
+//  void init() {}
+//}
