@@ -1,9 +1,10 @@
 import 'package:flutter/cupertino.dart';
+import 'package:osam/domain/state/base_state.dart';
 import 'package:osam/domain/store/store.dart';
 import 'package:provider/provider.dart';
 
-class StoreProvider extends StatelessWidget {
-  final Store store;
+class StoreProvider<ST extends BaseState> extends StatelessWidget {
+  final Store<ST> store;
   final Widget child;
   const StoreProvider({Key key, this.store, this.child}) : super(key: key);
 
@@ -15,5 +16,6 @@ class StoreProvider extends StatelessWidget {
     );
   }
 
-  static Store of(BuildContext context) => Provider.of<Store>(context);
+  static Store<ST> of<ST extends BaseState>(BuildContext context) =>
+      Provider.of<Store<ST>>(context);
 }
